Initial commit of the homie-benq-projector-rs232 firmware
This commit serves as the base to progress the firmware. It implements a
very basic behavior which is allow to set power on and off with a
predictive (and quite dumb) status reporting. The implementation is as
simple as it gets and will change quite a bit as the project progresses.